+++ featuretools/primitives/standard/api.py
@@ -2,4 +2,5 @@
 from .aggregation_primitives import *
 from .binary_transform import *
 from .cum_transform_feature import *
+from .rolling_transform_primitive import *
 from .transform_primitive import *