+++ featuretools/tests/synthesis/test_deep_feature_synthesis.py
@@ -36,6 +36,11 @@ from featuretools.primitives import (
     NotEqual,
     NumCharacters,
     NumUnique,
+    RollingCount,
+    RollingMax,
+    RollingMean,
+    RollingMin,
+    RollingSTD,
     Sum,
     TimeSincePrevious,
     TransformPrimitive,
@@ -410,6 +415,37 @@ def test_bad_groupby_feature(es):
                              agg_primitives=['sum'],
                              trans_primitives=[],
                              groupby_trans_primitives=['max'])
+
+
+@pytest.mark.parametrize(
+    "rolling_primitive",
+    [
+        RollingMax,
+        RollingMean,
+        RollingMin,
+        RollingSTD,
+    ]
+)
+def test_make_rolling_features(rolling_primitive, pd_es):
+    rolling_primitive_obj = rolling_primitive(window_length=7, gap=3, min_periods=5)
+    dfs_obj = DeepFeatureSynthesis(target_dataframe_name='log',
+                                   entityset=pd_es,
+                                   agg_primitives=[],
+                                   trans_primitives=[rolling_primitive_obj])
+    features = dfs_obj.build_features()
+    rolling_transform_name = f"{rolling_primitive.name.upper()}(datetime, value_many_nans, window_length=7, gap=3, min_periods=5)"
+    assert feature_with_name(features, rolling_transform_name)
+
+
+def test_make_rolling_count_off_datetime_feature(pd_es):
+    rolling_count = RollingCount(window_length=5, min_periods=3)
+    dfs_obj = DeepFeatureSynthesis(target_dataframe_name='log',
+                                   entityset=pd_es,
+                                   agg_primitives=[],
+                                   trans_primitives=[rolling_count])
+    features = dfs_obj.build_features()
+    rolling_transform_name = u"ROLLING_COUNT(datetime, window_length=5, min_periods=3)"
+    assert feature_with_name(features, rolling_transform_name)
 
 
 def test_abides_by_max_depth_param(es):